Kasutajaliideš

+++ LokaalneAndmepyydja/IAP.cs
@@ -7,7 +7,7 @@ using System.Numerics;
 using System.Runtime;
 
 using DatePriceT = System.Tuple<System.DateTime, float>;
-using VecT = System.Collections.Generic.List<System.Tuple<System.DateTime, float>>;
+using VecT = System.Collections.Generic.List<System.Tuple<System.DateTime, double>>;
 
 namespace Andmepyydja
 {
